requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
docopt==0.6.2
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Js2Py==0.74
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.1
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
ollama==0.6.1
openai-whisper==20250625
packaging==26.0
pipwin==0.5.2
PyAudio==0.2.14
pydantic==2.12.5
pydantic_core==2.41.5
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
regex==2026.4.4
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
SpeechRecognition==3.16.0
starlette==1.0.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
